+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next version
 
-- Included new parameters for `express-session` updates made to Roosevelt.
+- Included new parameters for `express-session` and `csrf-csrf` updates made to Roosevelt.
 - Added the ability to define the secrets directory name.
 - Fixed a bug that displayed `undefined` as the directory name on the app start instructions.